implement RTree insert

+++ src/BoundingBox.hs
@@ -4,7 +4,7 @@ module BoundingBox where
 data BoundingBox = BoundingBox {x1 :: !Double, y1 :: !Double, x2 :: !Double, y2 :: !Double}
   deriving (Eq)
 
-type Point = (Float, Float)
+type Point = (Double, Double)
 
 --get the least enlargement boundingbox that contains bb1 and bb2
 enlarge :: BoundingBox -> BoundingBox -> BoundingBox
@@ -27,5 +27,5 @@ intersect :: BoundingBox -> BoundingBox -> Bool
 intersect (BoundingBox x1 y1 x2 y2) (BoundingBox x1' y1' x2' y2') = not (x1 >= x2' || x2 <= x1' || y2 <= y1' || y2' <= y1)
 
 instance Show BoundingBox where
-	show (BoundingBox x1 y1 x2 y2) = "BoundingBox [" ++ show x1 ++ "," ++ show y1 ++ "," ++ show x2 ++ "," ++ show y2 ++ "]"
+    show (BoundingBox x1 y1 x2 y2) = "BoundingBox [" ++ show x1 ++ "," ++ show y1 ++ "," ++ show x2 ++ "," ++ show y2 ++ "]"
